# public/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get a page of reviews",
                "operationId": "dada4e5ef2019eed08806b379a8cfaa7",
                "requestBody": {
                    "description": "Page data",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "page": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reviews retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Review"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total_pages": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Out of range",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Out of range"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a new review",
                "operationId": "dded645e3969483bb1ef716b0f29dde7",
                "requestBody": {
                    "description": "Review data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewApiFormType"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully added review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully added review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Error"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Movie not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Review already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reviews/{id}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get a review by its ID",
                "operationId": "37bd43cc2432b16eef1d877b9216989f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the review",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update a review",
                "operationId": "08b5dd1e32f46ab3a8793154627f6d4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the review",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Review data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewApiFormType"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully updated review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Error"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cannot change another user's review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cannot change another user's review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Review or movie not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review or movie not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review",
                "operationId": "be9e3e830599ae67b7a53b3b76572810",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the review",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully deleted review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cannot delete another user's review",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cannot delete another user's review"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/reviews/{id}/votes": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get the votes of a review",
                "operationId": "9e26a877eb9f348dd662979f9aa4351e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the review",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Votes retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vote"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Review not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Movie": {
                "title": "Movie model",
                "description": "Movie model",
                "properties": {
                    "id": {
                        "description": "ID of the movie",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "title": {
                        "description": "Title of the movie",
                        "type": "string",
                        "example": "The Godfather"
                    },
                    "running_time": {
                        "description": "Running time of the movie",
                        "type": "integer",
                        "example": 175
                    },
                    "cover_photo": {
                        "description": "Cover photo of the movie",
                        "type": "string",
                        "example": "path/to/image.jpg"
                    },
                    "overview": {
                        "description": "Overview of the movie",
                        "type": "string",
                        "example": "The aging patriarch of an organized crime dynasty transfers control of his clandestine empire to his reluctant son."
                    },
                    "release_date": {
                        "description": "Release date of the movie",
                        "type": "string",
                        "example": "1972-03-14T00:00:00+00:00"
                    }
                },
                "type": "object"
            },
            "Review": {
                "title": "Review model",
                "description": "Review model",
                "properties": {
                    "id": {
                        "description": "ID of the review",
                        "type": "integer",
                        "format": "int64",
                        "example": 1,
                        "nullable": true
                    },
                    "movie": {
                        "$ref": "#/components/schemas/Movie"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "score": {
                        "description": "Score of the review",
                        "type": "integer",
                        "example": 5
                    },
                    "comment": {
                        "description": "Comment of the review",
                        "type": "string",
                        "example": "This movie was amazing!"
                    },
                    "date_reviewed": {
                        "description": "Date the review was written",
                        "type": "string",
                        "example": "2022-01-01T00:00:00+00:00"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User model",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "username": {
                        "description": "Username of the user",
                        "type": "string",
                        "example": "john_doe"
                    },
                    "restricted": {
                        "description": "Whether the user is restricted",
                        "type": "boolean",
                        "example": false
                    },
                    "date_joined": {
                        "description": "Date the user joined",
                        "type": "string",
                        "example": "2022-01-01T00:00:00+00:00"
                    },
                    "profile_image": {
                        "description": "Profile image of the user",
                        "type": "string",
                        "example": "path/to/image.jpg"
                    }
                },
                "type": "object"
            },
...
        }
    }
}